requirements.txt
Requests>=2.31.0
python-dotenv>=1.0.1
